imaging-big start
1: Implimented the start of some image making functions!  These use bytesIO to create a in-memory version of a image.

it's also going to be moved for display.py in main to imaging.py in imagemakingfunction.

Replaced targeting with target_maker.

Added new command.  getimage.  It pastes a PIL image into a channel.  RN that image is the example card.

Fixed Typo in target_maker.
Fixed a error in userprofile.